refactor(web): desacoplate component modal
- [x] remove sign-in rules from main file "index.tsx" and move to
  "SignInModal.tsx"
- [x] create new Modal to reuses in SignIn Modal and CreateNewHabit
  Modal
