+++ apps/web/src/app/components/Header.tsx
@@ -5,34 +5,26 @@ import habitLogo from '../../assets/habit-logo.svg'
 
 import { Plus, X } from 'phosphor-react'
 import { NewHabitForm } from './NewHabitForm'
+import * as Modal from './Modal'
 
 export function Header() {
   return (
     <div className="w-full max-w-3xl mx-auto flex items-center justify-between">
       <img src={habitLogo} alt="Habit Logo" />
 
-      <Dialog.Root>
-        <Dialog.Trigger className="border border-violet-500 font-semibold rounded-lg px-6 py-4 flex items-center gap-3 hover:border-violet-300 transition-colors focus:outline-none focus:ring-2 focus:ring-violet-700 focus:ring-offset-2 focus:ring-offset-[#09090A]">
+      <Modal.Box>
+        <Modal.ButtonTrigger>
           <Plus size={20} className="text-violet-500" />
           Novo hábito
-        </Dialog.Trigger>
+        </Modal.ButtonTrigger>
 
-        <Dialog.Portal>
-          <Dialog.Overlay className="w-screen h-screen bg-black/80 fixed inset-0" />
-
-          <Dialog.Content className="absolute p-10 bg-zinc-900 rounded-2xl w-full max-w-md top-1/2 left-1/2 -translate-x-1/2 -translate-y-1/2 text-white">
-            <Dialog.Close className="absolute right-6 top-6 text-zinc-400 hover:text-zinc-200 rounded-lg focus:outline-none focus:ring-2 focus:ring-violet-600 focus:ring-offset-2 focus:ring-offset-[#09090A]">
-              <X size={24} aria-label="Fechar" />
-            </Dialog.Close>
-
-            <Dialog.Title className="text-3xl leading-tight font-extrabold">
-              Criar hábito
-            </Dialog.Title>
-
-            <NewHabitForm />
-          </Dialog.Content>
-        </Dialog.Portal>
-      </Dialog.Root>
+        <Modal.Main
+          closeModalButton={<X size={24} aria-label="Fechar" />}
+          title="Criar hábito"
+        >
+          <NewHabitForm />
+        </Modal.Main>
+      </Modal.Box>
     </div>
   )
 }
